# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "StudyMate Learning Platform API",
    "description": "Production API specification for StudyMate v1.0 — Universal AI-Powered Learning, Academic Intelligence, and Institutional Management Platform.",
    "version": "1.0.0",
    "contact": {
      "name": "StudyMate Engineering Team",
      "email": "engineering@studymate.ai",
      "url": "https://studymate.ai"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development / Container Server"
    },
    {
      "url": "https://api.studymate.ai",
      "description": "Production Cloud Run Cluster"
    }
  ],
...
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register New User Account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserAccount" }
              }
            }
          },
          "400": { "description": "Invalid input or username already taken" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Authenticate User Credentials",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserAccount" }
              }
            }
          },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RegisterRequest": {
        "type": "object",
        "required": ["username", "email", "password"],
        "properties": {
          "username": { "type": "string", "example": "alex_student" },
          "email": { "type": "string", "example": "alex@university.edu" },
          "password": { "type": "string", "example": "SecurePass123!" },
          "displayName": { "type": "string", "example": "Alex Student" },
          "role": { "type": "string", "enum": ["student", "lecturer", "institution_admin"] },
          "subscription": { "type": "string", "enum": ["free", "premium", "institution"] }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string", "example": "alex_student" },
          "password": { "type": "string", "example": "SecurePass123!" }
        }
      },
      "UserAccount": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "username": { "type": "string" },
          "email": { "type": "string" },
          "displayName": { "type": "string" },
          "role": { "type": "string" },
          "subscription": { "type": "string" }
        }
      },
...
    }
  }
}